requirements.txt
feedparser>=6.0.10
openai>=1.0.0
lark-oapi>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pyyaml>=6.0
python-dateutil>=2.8.2
